Fix + change service distance and angle types to double/float64

+++ ros_workspace/src/manager/src/client/ActionClient.hpp
@@ -14,7 +14,7 @@
 #include <iostream>
 #include "order_codes.hpp"
 
-class ActionClient : public ClientT<action_msg_srv::srv::Order, action_msg_srv::srv::Order::Request, int64_t, int64_t, int64_t, int64_t> {
+class ActionClient : public ClientT<action_msg_srv::srv::Order, action_msg_srv::srv::Order::Request, int64_t, double, int64_t, double> {
 public:
     ActionClient() : ClientT("action") {};
 };
